requirements.txt
gunicorn
flask
flask_cors
pymongo
dnspython
pandas
python-dotenv